% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_chunk.R
\name{annotation_chunk}
\alias{annotation_chunk}
\title{Annotation chunk}
\usage{
annotation_chunk(text)
}
\arguments{
\item{text}{text to be displayed.}
}
\description{
An annotation chunk.
}
\examples{
annotation_chunk("This is a comment on some results.")
}
