% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{commands_chunk}
\alias{commands_chunk}
\title{Commands chunk}
\usage{
commands_chunk(commands.filename, stdout.filename = NULL,
  stderr.filename = NULL, title = "", description = "",
  collapsed = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Displays commands executed.
}
