% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{equal_distribution_model}
\alias{equal_distribution_model}
\title{equal_distribution_model}
\usage{
equal_distribution_model(alphabet, length)
}
\arguments{
\item{alphabet}{The alphabet the sequence should be based on as a vector}

\item{length}{The length of the generated sequence as an integer}
}
\value{
A vector that contains the letters of the generated sequence
}
\description{
Generate a random sequence in the equal distribution model with a specific length over a specified alphabet.
The probability \eqn{\mathbb{P}(\textbf{s})} of a sequence \eqn{\textbf{s}} over an alphabet \eqn{\Sigma} with length \eqn{n} is calculated by
\eqn{\mathbb{P}(\textbf{s}) = \frac{1}{|\Sigma|^n}}
}
