% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitous.R
\name{extract_parameters}
\alias{extract_parameters}
\title{Extract parameters}
\usage{
extract_parameters(include.wd = FALSE)
}
\arguments{
\item{include.wd}{Add working directory to parameter list.}
}
\value{
A list of the function's parameters and values.
}
\description{
Helper function to make a named list of function parameters and values.
}
\examples{
my_func <- function(x, a=2, ...) {
   extract_parameters()
}
my_func(10, b=4)
}
