% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_chunk.R
\name{figure_chunk}
\alias{figure_chunk}
\title{Figure chunk}
\usage{
figure_chunk(fun, title = "", description = "", width = 6, height = 5,
  pdf.filename = NULL, collapsed = FALSE)
}
\arguments{
\item{fun}{function to generate the figure.}

\item{title}{the title of the figure.}

\item{description}{description (or caption) of the figure to be displayed right below the title.}

\item{width}{width of the figure in inches}

\item{height}{height of the figure in inches}

\item{pdf.filename}{If specified, a PDF will be automatically generated and linked from the main document.}
}
\description{
A figure chunk displays a figure.
}
