% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{gc_content}
\alias{gc_content}
\title{gc_content}
\usage{
gc_content(sequence, frame_length, interpretation = TRUE)
}
\arguments{
\item{sequence}{The sequence to be analysed}

\item{frame_length}{The length of the individual analysing frames}

\item{interpretation}{A switch to determine if a vector with the relative frequencies should be returned or a value of how much the GC contents deviate from
a random distribution.}
}
\value{
Either a vector of relative GC contents or a single number indicating the deviation of a random GC distribution based on the value of
\code{interpretation}
}
\description{
This proceedure operates in the exact same way as \code{\link{gc_plot}} however instead of graphing the results of the analysis it returns its
results. It is slower than \code{\link{gc_plot}} since it calculates all relative frequencies. If the deviation result is close to zero the GC distribution
is probably random. If it's closer to one the distribution is probably not random.
}
