% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{gc_plot}
\alias{gc_plot}
\title{gc_plot}
\usage{
gc_plot(sequence, frame_length, plot_resolution = 500)
}
\arguments{
\item{sequence}{The sequence to be analysed}

\item{frame_length}{The length of the individual analysing frames}

\item{plot_resolution}{The number of how much frequencies should be displayed}
}
\description{
Calculate the GC content of a given sequence in frames with a given length and plot the GC content as a function of the \eqn{i}-th frame.
The way this is calculated is described in \code{\link{rel_freq}}. By viewing the plot it is possible to estimate if GC contents are random or not. Given
the relative frequencies \eqn{f} of a variable \eqn{x} in a frame with length \eqn{l} the true and unknown probability of \eqn{x} \eqn{p} can be described as
an interval \eqn{p \in \left[f - \frac{2\sigma}{\sqrt{l}}, f + \frac{2\sigma}{\sqrt{l}} \right]} wheras \eqn{\sigma} is the standard deviation of \eqn{x}.
The interval describes the \eqn{95\%} confidence interval thus in \eqn{5\%} of all cases the true value of \eqn{p} is element of the interval. Since the true
standard diviation \eqn{\sigma = \sqrt(p(1-p))} cannot be calculated it has to be estimated by \eqn{\sigma = \sqrt{f(1 - f)}}. If used on the GC contents the
random probability is given by \eqn{p = \frac{1}{4} + \frac{1}{4} = \frac{1}{2}} and thus the interval
\eqn{\left[\frac{1}{2} - \frac{1}{\sqrt{l}}, \frac{1}{2} + \frac{1}{\sqrt{l}}\right]} which is also plotted. So if most of the GC contents are not in the
plotted interval a non random GC content can be assumed.
}
