% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitous.R
\name{gen_id}
\alias{gen_id}
\title{Generate a sequencial ID.}
\usage{
gen_id(type = "global")
}
\value{
An integer ID number guaranteed to by different from all previously generated IDs.
}
\description{
Generate a sequencial ID.
}
