% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{initialize_markow}
\alias{initialize_markow}
\title{initialize_markow}
\usage{
initialize_markow(sequence, alphabet, round = 2)
}
\arguments{
\item{sequence}{The sequence the markow model should be based on}

\item{alphabet}{The alphabet that is used by the sequence}

\item{round}{The number of digits the contents of the transition matrix should be rounded to. If this value is zero or lower the values will stay as calculated.}
}
\value{
The calculated transition matrix
}
\description{
Calculates the transition matrix assuming a markow model specified by a given sequence with the respective alphabet. The exact calculation
is described in \link{rel_freq}
}
