% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{initialize_multinomial}
\alias{initialize_multinomial}
\title{initialize_markow}
\usage{
initialize_multinomial(sequence, alphabet, round = 2)
}
\arguments{
\item{sequence}{The sequence the multinomial distribution model should be based on}

\item{alphabet}{The alphabet that is used by the sequence}

\item{round}{The number of digits the contents of the probability vector should be rounded to. If this value is zero or lower the values will stay as calculated.}
}
\value{
The calculated probability vector
}
\description{
Calculates the probability vector assuming a multinomial distribution model specified by a given sequence with the respective alphabet.
The exact calculation is described in \link{rel_freq}
}
