% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{markow_model}
\alias{markow_model}
\title{markow_model}
\usage{
markow_model(alphabet, length, matrix, initial_probabilities = NULL)
}
\arguments{
\item{alphabet}{The alphabet the sequence should be based on as a vector}

\item{length}{The length of the generated sequence as an integer}

\item{matrix}{The transition matrix so that \code{matrix[i, j]} describes the probability that a state \eqn{i} is followed by state \eqn{j}}

\item{initial_probabilities}{A vector of probabilities that a sequence is starting with the \eqn{i}-th letter of the alphabet. If this argument is empty
an equal distribution is assumed (which does not matter for large sequences)}
}
\value{
A vector that contains the letters of the generated sequence. If either the sum of the rows of the transition matrix or the sum of the initial probabilities
are not equal to 1 a warning is thrown
}
\description{
Generate a random sequence in the markow model with a specific lenght over a specified alphabet
with a specified transition matrix and initial probabilities. Given an index set \eqn{I} with \eqn{r} elements then a \eqn{r\times r} matrix \eqn{(p_{ij})_{i,j \in I}}
of nonnegative numbers is a transition matrix if \eqn{\forall i \in I: \sum_{j \in I} p_{ij} = 1}. \eqn{p_{ij}} is the probability that a state \eqn{i} is followed
by state \eqn{j}. Given an Alphabet \eqn{\Sigma}, nonnegative numbers \eqn{(p_i)_{i \in \Sigma}} with sum \eqn{1} and a transition matrix \eqn{P = (p_{ij})_{i,j \in \Sigma}}
the probability \eqn{\mathbb{P}(\textbf{s})} of a sequence \eqn{(\textbf{s})} over \eqn{\Sigma} with length \eqn{n} is calculated by
\eqn{\mathbb{P}(\textbf{s}) = p_{s_1}\prod_{l = 2}^n p_{s_{n-1}s_n}}
}
