% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{multinomial_distribution_model}
\alias{multinomial_distribution_model}
\title{multinomial_distribution_model}
\usage{
multinomial_distribution_model(alphabet, length, probabilities)
}
\arguments{
\item{alphabet}{The alphabet the sequence should be based on as a vector}

\item{length}{The length of the generated sequence as an integer}

\item{probabilities}{A vector of probabilities so that the \eqn{i}-th element is the probability of the \eqn{i}-th letter of the alphabet.}
}
\value{
A vector that contains the letters of the generated sequence
}
\description{
Generate a random sequence in the multinomial distribution model with a specific length over a specified alphabet
with specified probabilities. Given an alphabet \eqn{\Sigma} and nonnegative Numbers \eqn{(p_i)_{i \in \Sigma}} with sum \eqn{1}
(\eqn{\forall i \in \Sigma: p_i \in [0, 1]}) the probability \eqn{\mathbb{P}(\textbf{s})} of a sequence \eqn{\textbf{s}} over \eqn{\Sigma} with length \eqn{n}
is calculated by \eqn{\mathbb{P}(\textbf{s}) = \prod_{l = 1}^{n} p_{s_l}}
}
