% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{read_dna_sequence}
\alias{read_dna_sequence}
\title{read_dna_sequence}
\usage{
read_dna_sequence(path, filetype)
}
\arguments{
\item{path}{The path to the file to be read}

\item{filetype}{The type of the file to be read currently supported are the values \code{"fasta"} for .fasta files and \code{"dat"} for .dat files}
}
\value{
The DNA sequence as a vector of characters
}
\description{
Read a file of specified filetype containing DNA information and return the DNA sequence as a vector of characters.
}
