% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{rel_freq}
\alias{rel_freq}
\title{rel_freq}
\usage{
rel_freq(sequence, alphabet, wordsize = 1, start = 0, count = FALSE)
}
\arguments{
\item{sequence}{The sequence to be analysed}

\item{alphabet}{The underlying alphabet of the input sequence}

\item{wordsize}{The size of the words to be analysed}

\item{start}{The start of the sequence}

\item{count}{Switch if absolute instead of relative frequencies are desired}
}
\value{
A vector containing the absolute or relative frequencies of the possible words with length \code{wordsize} found in the input sequence
}
\description{
Get the relative frequencies of words (pairs) in a given sequence with a specified alphabet.
This can also used as a approximation for the required probabilities for the markow model or the multinomial distribution model. The relative frequency
\eqn{\hat{p_i}} for a given state \eqn{i} in a sequence \eqn{\textbf{s}_{[1, L]}} with length \eqn{L}
is given by \eqn{\hat{p_i} = \frac{\text{\# } i \text{ in }\textbf{s}_{[1, L]}}{L}} whilst the relative frequency of \eqn{\hat{p_{ij}}} for
a given state \eqn{i} that is followed by state \eqn{j} in \eqn{\textbf{s}} is determined by
\eqn{\hat{p_{ij}} = \frac{\text{\# } ij \text{ in } \textbf{s}_{[1, L]}}{\text{\# } i \text{ in }\textbf{s}_{[1, L - 1]}}}. If the single \eqn{\hat{p_i}} are
clearly differing than an equal distribution can be ruled out.
If the \eqn{\hat{p_{ij}}} are clearly differing from the single \eqn{\hat{p_j}} a multinomial distribution can be ruled out.
}
