% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_chunk.R
\name{render_annotation_chunk}
\alias{render_annotation_chunk}
\title{Default function to render an annotation chunk.}
\usage{
render_annotation_chunk(chunk, sec_path, ...)
}
\arguments{
\item{chunk}{chunk definition.}

\item{sec_path}{vector of section numbers.}

\item{header}{whether to treat title as a subsection header (and display it in the table of contents).}
}
\description{
Default function to render an annotation chunk.
}
