% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitous.R
\name{render_chunk}
\alias{render_chunk}
\title{Render a chunk.}
\usage{
render_chunk(chunk)
}
\arguments{
\item{chunk}{}
}
\description{
TODO: make this output a structure with a print handler
right now one has to "cat" the results
ideally, in the future, figure_chunk should print itself
}
