% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitous.R
\name{render_document}
\alias{render_document}
\title{Render document.}
\usage{
render_document(document, filename, chunk_functions_ex = NULL,
  section_functions = NULL, title = "Untitled", theme = "default")
}
\arguments{
\item{document}{a section, defining the document structure.}

\item{filename}{the target filename.}

\item{chunk_functions_ex}{functions used to render chunks (see details).}

\item{section_functions}{functions used to render sections (a vector, defining the function to use at each section depth).}
}
\description{
This function renders the html document.
}
