% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_chunk.R
\name{render_figure_chunk}
\alias{render_figure_chunk}
\title{Default function to render a figure chunk.}
\usage{
render_figure_chunk(chunk, sec_path, header = TRUE, ...)
}
\arguments{
\item{chunk}{chunk definition.}

\item{sec_path}{vector of section numbers.}

\item{header}{whether to treat title as a subsection header (and display it in the table of contents).}
}
\description{
Default function to render a figure chunk.
}
