% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{render_script_chunk}
\alias{render_script_chunk}
\title{Default function to render a script chunk.}
\usage{
render_script_chunk(chunk, header = TRUE, ...)
}
\arguments{
\item{chunk}{chunk definition.}

\item{header}{whether to treat title as a subsection header (and display it in the table of contents).}

\item{sec_path}{vector of section numbers.}
}
\description{
Default function to render a script chunk.
}
