% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{render_section_basic}
\alias{render_section_basic}
\title{A basic section style.}
\usage{
render_section_basic(ns, depth, title, description, parameters, body)
}
\arguments{
\item{body}{}
}
\description{
This is an internal function not meant to be called by the user.
}
\examples{
<div>',
</div>',
}
