% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{render_section_panel3}
\alias{render_section_panel3}
\title{A nice panel style for sections.}
\usage{
render_section_panel3(ns, depth, title, description, parameters, body)
}
\arguments{
\item{body}{}
}
\description{
A nice panel style for sections.
}
