% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_chunk.R
\name{render_table_chunk}
\alias{render_table_chunk}
\title{Default function to render a table chunk.}
\usage{
render_table_chunk(chunk, sec_path, header = TRUE, ...)
}
\arguments{
\item{chunk}{chunk definition.}

\item{sec_path}{vector of section numbers.}

\item{header}{whether to treat title as a subsection header (and display it in the table of contents).}
}
\description{
Default function to render a table chunk.
}
