% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs_chunk.R
\name{render_tabs_chunk}
\alias{render_tabs_chunk}
\title{Default function to render a tabs chunk.}
\usage{
render_tabs_chunk(chunk, sec_path, params, ...)
}
\arguments{
\item{chunk}{chunk definition.}

\item{sec_path}{vector of section numbers.}

\item{...}{}

\item{header}{whether to treat title as a subsection header (and display it in the table of contents).}
}
\description{
Default function to render a tabs chunk.
}
