% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubiquitous.R
\name{rf_plot}
\alias{rf_plot}
\title{rf_plot}
\usage{
rf_plot(
  rfs,
  sequence,
  max_rf_length = 300,
  quantile = 0.95,
  show_orf_quantile = TRUE
)
}
\arguments{
\item{rfs}{Input reading frames/open reading frames}

\item{max_rf_length}{The maximum length of the reading frames}

\item{quantile}{The level of confidence needed for quantile calculation}

\item{show_orf_quantile}{Plot the quantile}
}
\value{
The lowest probably (confidence level) not randomly distributed open reading frame length
}
\description{
Plot the frequencies of the width of the input reading frames in a histogram. Assuming the original sequence was a multinomially distributed random
DNA Sequence a length can be calculated so that an open reading frame greater than that length is probably not described by a random distribution and thus
biologically relevant. This length can be approximated by a quantile with a specified level of confidence. The quantile is also plotted. The probability that
an open reading frame with length \eqn{\geq k} is found in a random distribution can be calculated as
\eqn{\mathbb{P}(k) = \mathbb{P}(\text{STOP}) \cdot (1 - \mathbb{P}(\text{STOP})^{k-1}) = \frac{3}{64} \cdot \left(1 - \frac{3}{64}\right)^{k - 1}} which
can also easily be calculated by \code{1 - ecdf(width(orfs))(k - 1)}. This equals to the p-value of how random the open reading frame actually is.
}
