% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{run_system}
\alias{run_system}
\title{Run shell commands with logging}
\usage{
run_system(commands, log.basename, append = TRUE)
}
\arguments{
\item{append}{}
}
\description{
Run shell commands with logging
}
