% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{run_system_parallel}
\alias{run_system_parallel}
\title{Run shell commands with logging}
\usage{
run_system_parallel(commands, log.basename, append = TRUE, max.cores = 8)
}
\arguments{
\item{max.cores}{}
}
\description{
Run shell commands with logging
}
