% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{script_chunk}
\alias{script_chunk}
\title{Script chunk}
\usage{
script_chunk(commands, title = "", description = "", collapsed = FALSE,
  ...)
}
\arguments{
\item{...}{}
}
\description{
A script chunk runs a script.
}
