% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{section}
\alias{section}
\title{Define a section.}
\usage{
section(title, chunks = NULL, subsections = NULL, par = NULL,
  description = "")
}
\arguments{
\item{title}{the title of the section.}

\item{chunks}{list of chunks to display.}

\item{subsections}{list of subsections.}

\item{par}{optional list of parameters to display in header.}

\item{description}{description of the section, to display below the title.}
}
\description{
A section has a title, chunks to display content, and may contain subsections.
}
