% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/section.R
\name{source_section}
\alias{source_section}
\title{Source a section from a file.}
\usage{
source_section(title, sourcename, funname, params = list(),
  description = NULL)
}
\arguments{
\item{title}{the title of the section.}

\item{sourcename}{the file containing the section function.}

\item{funname}{name of the function to call.}

\item{params}{parameters to pass to the function}

\item{description}{description of the section (optional). If not NULL will replace the description provided by the function.}
}
\description{
Convenience function to create a section by sourcing an script.
}
