% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_chunk.R
\name{table_chunk}
\alias{table_chunk}
\title{Table chunk}
\usage{
table_chunk(dataframe, title = "", description = "", scrollX = FALSE)
}
\arguments{
\item{dataframe}{the table to display.}

\item{title}{the title of the table.}

\item{description}{description (or caption) of the table to be displayed right below the title.}

\item{scrollX}{whether add a horizontal scrollbar to the table}
}
\description{
A table chunk displays a table. By default, the table will be generated using \code{\link{kable}} and \code{\link{kable_styling}}.
}
