% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk.R
\name{table_link}
\alias{table_link}
\title{Make an HTML link to a table.}
\usage{
table_link(x, text = "Table", filename = NULL)
}
\arguments{
\item{x}{a dataframe or a character string indicating a file name.}

\item{text}{}
}
\description{
Make an HTML link to a table.
}
\examples{
# embeds iris encoded in base64 and makes a link
table_link(iris, text="Iris")

# saves iris to a file (as tab-separated values) and makes a link
table_link(iris, text="Iris", filename="html_files/iris.tsv")

# includes a link to my_table.tsv
table_link("my_table.tsv", text="My table")
}
