% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabs_chunk.R
\name{tabs_chunk}
\alias{tabs_chunk}
\title{Tabs chunk.}
\usage{
tabs_chunk(tabchunks, tabnames, title = "", description = "")
}
\arguments{
\item{description}{}
}
\description{
The tabs chunk displays other chunks inside tabs.
}
