% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_chunk.R
\name{text_chunk}
\alias{text_chunk}
\title{Text chunk}
\usage{
text_chunk(text)
}
\arguments{
\item{text}{text to be displayed.}
}
\description{
A text chunk displays some text.
}
\examples{
text_chunk("Hello there.")
text_chunk("A bit of *text* with some **markdown**")
}
