% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitous.R
\docType{package}
\name{UBIquitous-package}
\alias{UBIquitous-package}
\title{UBIquitous: A Package for Generation of Structured Reports}
\description{
This package allows the user to create HTML reports by specifying a
document structure consisting of sections and *chunks*. It includes
modules to process NGS datasets.
}
\section{Chunks}{

Chunks are displayable elements in a report (e.g. a table, a figure, a paragraph).

\code{\link{text_chunk}} specifies a piece of text.
\code{\link{table_chunk}} specifies a table.
\code{\link{figure_chunk}} specifies a figure.
}

\examples{
iris_section <- function(num.rows, colors) {
  tab1 <- table_chunk(head(iris, num.rows),
                      title = "The iris dataset",
                      description = paste("This table displays the first", num.rows, "rows of the iris dataset."))

  fig1 <- figure_chunk(fun = function() {
     pairs(iris)
  }, title = "Pairs plot of iris", description = "")

  section <- list(
     title = "Iris",
     description = "A section about the iris dataset.",
     chunks = list(tab1, fig1),
     par = extract_parameters()
  )
}

sec1 <- iris_section(6, colors = c("red", "green", "blue"))
document <- list(sections=list(sec1))
render_document(document, output.dir = ".", output.file = "Example1.html")

}
\author{
Daniel Neves, \email{danielneves2005@gmail.com}
}
