% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UBIquitous.R
\name{render_section}
\alias{render_section}
\title{Render a single section}
\usage{
render_section(section, title = "Untitled", depth = 1,
  render_function = render_section_basic)
}
\arguments{
\item{section}{a section (may contain subsections)}

\item{chunk_functions_ex}{functions used to render chunks (see details).}

\item{section_functions}{functions used to render sections (a vector, defining the function to use at each section depth).}
}
\description{
This function renders a section's HTML.
}
